requirements.txt
Django==4.2.2
djangorestframework==3.14.0
matplotlib==3.7.1
numpy==1.24.3
pandas==2.1.2
scikit_learn==1.3.1
seaborn==0.13.0
